+++ .github/workflows/release-rust.yml
@@ -55,7 +55,7 @@ jobs:
           if [ "$RUNNER_OS" == "Linux" ]; then
            sudo apt-get update -y
            sudo apt install -y libunwind-dev
-           sudo apt-get install -y libx11-dev nasm libdbus-1-dev libudev-dev libxcb-randr0-dev libgstreamer1.0-dev libgstreamer-plugins-base1.0-dev
+           sudo apt-get install -y libx11-dev nasm libdbus-1-dev libudev-dev libxcb-randr0-dev libgstreamer1.0-dev libgstreamer-plugins-base1.0-dev libxi-dev libxtst-dev
           elif [ "$RUNNER_OS" == "macOS" ]; then
            brew install nasm
           elif [ "$RUNNER_OS" == "Windows" ]; then
