+++ README.md
@@ -175,7 +175,7 @@ This program has been tested to work on:
 #### Ubuntu
 
 ```sh
-sudo apt-get install -y libx11-dev nasm libdbus-1-dev libudev-dev libxcb-randr0-dev libgstreamer1.0-dev libgstreamer-plugins-base1.0-dev
+sudo apt-get install -y libx11-dev nasm libdbus-1-dev libudev-dev libxcb-randr0-dev libgstreamer1.0-dev libgstreamer-plugins-base1.0-dev libxi-dev libxtst-dev
 ```
 
 ### Using `cargo-make`
