+++ app/src/persist.rs
@@ -1,7 +1,8 @@
-use std::{fs, path::PathBuf};
+use std::{fs, path::Path};
 
-use chrono::{serde::ts_seconds, Utc};
+use chrono::{serde::ts_seconds, Duration, Utc};
 use serde::{Deserialize, Serialize};
+use serde_json::Value;
 
 use crate::{profile::Profile, util::StorageTrait};
 
@@ -12,7 +13,8 @@ pub struct Settings {
 }
 
 impl Settings {
-    pub fn load_or_default(path: &PathBuf) -> Self {
+    /// Load the settings from the specified path or generate default ones if an error occurs
+    pub fn load_or_default(path: &Path) -> Self {
         let mut persist: Self = Self::default();
 
         if let Ok(string) = fs::read_to_string(path) {
@@ -20,6 +22,55 @@ impl Settings {
         }
 
         persist
+    }
+
+    /// Same as `load_or_default` but it also performs an update check and updates values accordingly
+    pub fn load_with_check(path: &Path) -> Self {
+        let mut settings = Self::load_or_default(path);
+
+        let version_name = &mut settings.updates.version_name;
+
+        let time_since_last_check = Utc::now() - settings.updates.last_checked;
+
+        if settings.updates.check_for_updates && time_since_last_check > Duration::days(1) {
+            let client = reqwest::blocking::Client::builder()
+                .user_agent(format!("4JX/L5P-Keyboard-RGB, Ver {}", env!("CARGO_PKG_VERSION")))
+                .build()
+                .unwrap();
+
+            if let Ok(res) = client.get("https://api.github.com/repos/4JX/L5P-Keyboard-RGB/tags").send() {
+                let json: Value = res.json().unwrap();
+
+                if let Some(entry) = json.pointer("/0/name") {
+                    let mut name = entry.to_string().replace('\"', "");
+
+                    match version_name {
+                        Some(current_name) => {
+                            if settings.updates.skip_version && current_name != name.as_mut() {
+                                *current_name = name;
+                                settings.updates.skip_version = false;
+                            }
+                        }
+                        None => {
+                            *version_name = Some(name);
+                            settings.updates.skip_version = false;
+                        }
+                    }
+                }
+            };
+
+            settings.updates.last_checked = Utc::now();
+        }
+
+        if version_name.is_some() {
+            let n = version_name.as_ref().unwrap();
+
+            if n.is_empty() || n == concat!("v", env!("CARGO_PKG_VERSION")) {
+                *version_name = None;
+            }
+        }
+
+        settings
     }
 }
