+++ app/src/profile.rs
@@ -26,7 +26,7 @@ impl Default for KeyboardZone {
 
 type Zones = [KeyboardZone; 4];
 
-#[derive(Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq)]
 pub struct Profile {
     pub name: String,
     pub rgb_zones: Zones,
@@ -34,18 +34,6 @@ pub struct Profile {
     pub direction: Direction,
     pub speed: u8,
     pub brightness: u8,
-}
-
-// Primarily differentiated by name but the rest is left for fuckups
-impl PartialEq for Profile {
-    fn eq(&self, other: &Self) -> bool {
-        self.name == other.name
-            && self.rgb_zones == other.rgb_zones
-            && self.effect == other.effect
-            && self.direction == other.direction
-            && self.speed == other.speed
-            && self.brightness == other.brightness
-    }
 }
 
 impl Default for Profile {
